update avatar done

+++ Backend/apps/auth/src/users/upload.controller.ts
@@ -4,12 +4,16 @@ import {
   Delete,
   Param,
   Post,
+  Req,
   UploadedFile,
+  UseGuards,
   UseInterceptors,
 } from '@nestjs/common';
 import { FileInterceptor } from '@nestjs/platform-express';
 import { UsersService } from './users.service';
 import { FilesService } from '../files/files.service';
+import { JwtAccessGuard } from '../auth/guards/jwt-access.guards';
+import { GraphQLError } from 'graphql';
 
 @Controller()
 export class UserHttpController {
@@ -18,11 +22,20 @@ export class UserHttpController {
     private readonly fileService: FilesService,
   ) {}
   @Post('upload-image')
+  @UseGuards(JwtAccessGuard)
   @UseInterceptors(FileInterceptor('file'))
   async upload(
     @UploadedFile() file: Express.Multer.File,
     @Body('id') id: string,
+    @Req() request: Request,
   ) {
+    if (request['user'].userId != id) {
+      throw new GraphQLError('You can not update other avatar', {
+        extensions: {
+          errorCode: '5001-15',
+        },
+      });
+    }
     const containerName = 'fileupload';
     const upload = await this.fileService.uploadFileHttp(file, containerName);
     this.userService.saveUrl(id, upload, containerName);
@@ -30,7 +43,15 @@ export class UserHttpController {
   }
 
   @Delete('remove-image/:id')
-  async remove(@Param('id') id: string) {
+  @UseGuards(JwtAccessGuard)
+  async remove(@Param('id') id: string, @Req() request: Request) {
+    if (request['user'].userId != id) {
+      throw new GraphQLError('You delete other avatar', {
+        extensions: {
+          errorCode: '5001-16',
+        },
+      });
+    }
     const containerName = 'fileupload';
     const user = await this.userService.remove(id, containerName);
     return {
